스크립트 변경
BattleScene -> DohyunBattle
